add new function

+++ routes/feedTransaction/index.js
@@ -1,9 +1,10 @@
 const express = require('express');
 const router = express.Router();
 
-const { getAll, create, deleteById } = require('./controller');
+const { getAll, getByUserId, create, deleteById } = require('./controller');
 
 router.get('/', getAll);
+router.get('/:id', getByUserId);
 router.post('/', create);
 router.delete('/:id', deleteById);
